Move time utils to folder where it logically belongs to
